Update java files to use Maven structure
