+++ src/META-INF/MANIFEST.MF
@@ -1,0 +1,3 @@
+Manifest-Version: 1.0
+Main-Class: main.java.com.rw.legion.DefaultJob
+
