--- src/com/rw/legion/columncheck/ColumnChecker.java
@@ -1,27 +1,0 @@
-/*
- * Copyright (C) 2017 Republic Wireless
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.rw.legion.columncheck;
-
-/**
- * Defines an interface for Legion "column checkers." A column checker must have
- * a "validates" method which accepts the string of data in the input and checks
- * to make sure it is in the expected format. Should return a boolean.
- */
-
-public interface ColumnChecker {
-    public boolean validates(String str);
-}
