--- src/com/rw/legion/columncheck/StringChecker.java
@@ -1,33 +1,0 @@
-/*
- * Copyright (C) 2017 Republic Wireless
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.rw.legion.columncheck;
-
-import com.google.gson.JsonObject;
-
-/**
- * A column checker that validates a string is a string. Always returns true.
- */
-
-public class StringChecker implements ColumnChecker {
-    public StringChecker(JsonObject json) {
-        
-    }
-    
-    public boolean validates(String str) {
-        return true;
-    }
-}
